// frontend/src/lib/api/types.gen.ts
export type Package = {
    name: string;
    version: string;
    type: string;
    format?: string;
    platform?: string;
    architecture?: string;
    size?: string;
    created?: string;
    updated?: string;
    pulled?: string;
    description?: string;
    origin?: string;
    index?: string;
    manifest?: string;
};